requirements.txt
fastapi
uvicorn
pygithub
openai
python-dotenv
PyJWT
cryptography
requests
